# perf: use `phf` for more efficient lookup tables