--- a/rust_snuba/src/processors/mod.rs
+++ b/rust_snuba/src/processors/mod.rs
@@ -11,14 +11,11 @@ use rust_arroyo::backends::kafka::types::KafkaPayload;
 
 pub type ProcessingFunction = fn(KafkaPayload, KafkaMessageMetadata) -> anyhow::Result<InsertBatch>;
 
-pub fn get_processing_function(name: &str) -> Option<ProcessingFunction> {
-    match name {
-        "FunctionsMessageProcessor" => Some(functions::process_message),
-        "ProfilesMessageProcessor" => Some(profiles::process_message),
-        "QuerylogProcessor" => Some(querylog::process_message),
-        "ReplaysProcessor" => Some(replays::process_message),
-        "SpansMessageProcessor" => Some(spans::process_message),
-        "MetricsSummariesMessageProcessor" => Some(metrics_summaries::process_message),
-        _ => None,
-    }
-}
+pub static PROCESSING_FUNCTIONS: phf::Map<&'static str, ProcessingFunction> = phf::phf_map! {
+    "FunctionsMessageProcessor" => functions::process_message,
+    "ProfilesMessageProcessor" => profiles::process_message,
+    "QuerylogProcessor" => querylog::process_message,
+    "ReplaysProcessor" => replays::process_message,
+    "SpansMessageProcessor" => spans::process_message,
+    "MetricsSummariesMessageProcessor" => metrics_summaries::process_message,
+};