@untitaker 2023-12-27
i wouldn't do this -- this is not a hot function and I am skeptical that for tables of this size, match is slower than phf. benchmark delta seems to be within error range.